% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwatershed.R
\name{rwatershed}
\alias{rwatershed}
\title{An R implementation of a watershed segmentation}
\source{
This function is part of the following repository at \url{https://github.com/lochbika/Rwatershed}
}
\usage{
rwatershed(x)
}
\arguments{
\item{x}{input matrix}

\item{periodic}{Use periodioc boundary conditions? Default: TRUE.}
}
\value{
Returns a matrix of the same dimensions as x.
}
\description{
This function should not be called directly. Instead, use watershed.
}
