% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watershed.R
\name{watershed}
\alias{watershed}
\title{Watershed algorithm}
\source{
This function is part of the following repository at \url{https://github.com/lochbika/Rwatershed}
}
\usage{
watershed(x, method = "simple", periodic = FALSE)
}
\arguments{
\item{x}{input matrix}

\item{method}{Chose the method to use. Must be a character. Currently supported method is "simple".}

\item{periodic}{Use periodioc boundary conditions? Default: False.}
}
\value{
Returns a matrix of the same dimensions as x.
}
\description{
Performs the watershed algorithm on a matrix
}
